requirements.txt
streamlit>=1.25.0
xgboost>=1.7.0
pandas>=1.5.0
numpy>=1.23.0
scikit-learn>=1.1.0
plotly>=5.10.0
matplotlib>=3.5.0
supabase-py>=1.0.3
google-generativeai>=0.3.0
shap>=0.41.0
python-dotenv>=1.0.0 